added games board quantity

+++ src/main/java/com/example/gamemanagement/Controllers/Student/ReservationListController.java
@@ -1,6 +1,7 @@
 package com.example.gamemanagement.Controllers.Student;
 
 import com.example.gamemanagement.db.DBconnection;
+import com.example.gamemanagement.utils.Clients;
 import com.example.gamemanagement.utils.Reservation;
 import com.example.gamemanagement.utils.UserInfo;
 import javafx.animation.KeyFrame;
@@ -19,7 +20,6 @@ import java.sql.SQLException;
 import java.util.ResourceBundle;
 
 public class ReservationListController implements Initializable {
-    public TableColumn<Reservation, String> col_id;
     public TableColumn<Reservation, String> col_game_name;
     public TableColumn<Reservation, String> col_cancel_reservation;
     public Label error_label;
@@ -56,23 +56,35 @@ public class ReservationListController implements Initializable {
             table_reservation.setItems(list);
             col_cancel_reservation.setCellFactory(param -> new TableCell<>() {
                 private final Button cancel = new Button("Cancel");
-
+                {
+                    cancel.setStyle("-fx-background-color: red; -fx-text-fill: white; -fx-font-weight: bold; -fx-font-size: 10px; -fx-max-width: 100px");
+                }
 
                 {
                     cancel.setOnAction(event -> {
+
+                        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                        alert.setTitle("Delete Reservation");
+                        alert.setHeaderText("Delete Reservation");
+                        alert.setContentText("Are you sure you want to delete this reservation?");
                         Reservation reservation = getTableView().getItems().get(getIndex());
-                        try {
-                            Connection connection = DBconnection.getInstance().getConnection();
-                            PreparedStatement preparedStatement = connection.prepareStatement("DELETE FROM reservations WHERE id = ?");
-                            preparedStatement.setObject(1, reservation.getReservationId());
-                            preparedStatement.executeUpdate();
-                            error_label.setText("Reservation cancelled successfully");
-                            Timeline timeline = new Timeline(new KeyFrame(Duration.seconds(3), event1 -> error_label.setText("")));
-                            timeline.play();
-                            table();
-                        } catch (SQLException throwable) {
-                            throwable.printStackTrace();
-                        }
+                        alert.showAndWait().ifPresent(response -> {
+                            if (response == javafx.scene.control.ButtonType.OK) {
+                                try {
+                                    Connection connection = DBconnection.getInstance().getConnection();
+                                    PreparedStatement preparedStatement = connection.prepareStatement("DELETE FROM reservations WHERE id = ?");
+                                    preparedStatement.setObject(1, reservation.getReservationId());
+                                    preparedStatement.executeUpdate();
+                                    error_label.setText("Reservation cancelled successfully");
+                                    error_label.setStyle("-fx-text-fill: red");
+                                    Timeline timeline = new Timeline(new KeyFrame(Duration.seconds(3), event1 -> error_label.setText("")));
+                                    timeline.play();
+                                    table();
+                                } catch (SQLException throwable) {
+                                    throwable.printStackTrace();
+                                }
+                            }
+                        });
                     });
                 }
